+++ src/test/java/keeper/project/homepage/ApiControllerTestSetUp.java
@@ -29,7 +29,7 @@ import keeper.project.homepage.admin.service.library.BookManageService;
 import keeper.project.homepage.service.attendance.AttendanceService;
 import keeper.project.homepage.service.member.MemberHasCommentDislikeService;
 import keeper.project.homepage.service.member.MemberHasCommentLikeService;
-import keeper.project.homepage.service.posting.CommentService;
+import keeper.project.homepage.user.service.posting.CommentService;
 import keeper.project.homepage.service.sign.SignUpService;
 import keeper.project.homepage.user.service.member.MemberService;
 import org.junit.jupiter.api.BeforeEach;
